Graph network block

Graph network (GN) is a more generic model for graph neural network. For details, a graph network block is defined as follow:

\[\begin{aligned} e_{ij}^{\prime} &= \phi^{e}(e_{ij}, v_i, v_j, u) \\ v_{i}^{\prime} &= \phi^{v}(\bar{e}_{i}^{\prime}, v_i, u) \\ u^{\prime} &= \phi^{u}(\bar{e}^{\prime}, \bar{v}^{\prime}, u) \end{aligned} \ \ \ \ \begin{aligned} \bar{e}_{i}^{\prime} &= \rho^{e \rightarrow v}(E_i^{\prime}) \\ \bar{e}^{\prime} &= \rho^{e \rightarrow u}(E^{\prime}) \\ \bar{v}^{\prime} &= \rho^{v \rightarrow u}(V^{\prime}) \end{aligned}\]

where $\phi$ and $\rho$ denote update functions and aggregate functions, respectively. $v_i$ and $v_j$ are node features from node $i$ and its neighbor node $j$, $e_{ij}$ is edge feature for edge $(i, j)$, and $u$ is global feature for whole graph. $e_{ij}^{\prime}$, $v_{i}^{\prime}$ and $u^{\prime}$ are new features for edge, node and global. $\bar{e}_{i}^{\prime}$, $\bar{e}^{\prime}$ and $\bar{v}^{\prime}$ are aggregated new features for edge, node and global.

Reference: Peter W. Battaglia, Jessica B. Hamrick, Victor Bapst, Alvaro Sanchez-Gonzalez, Vinicius Zambaldi, Mateusz Malinowski, Andrea Tacchetti, David Raposo, Adam Santoro, Ryan Faulkner, Caglar Gulcehre, Francis Song, Andrew Ballard, Justin Gilmer, George Dahl, Ashish Vaswani, Kelsey Allen, Charles Nash, Victoria Langston, Chris Dyer, Nicolas Heess, Daan Wierstra, Pushmeet Kohli, Matt Botvinick, Oriol Vinyals, Yujia Li, Razvan Pascanu (2018)

Ingraph network, it describes an update order: edge, node and then global. There are three corresponding update functions for edge, node and then global, respectively. Three update functions return their default values as follow:

update_edge(gn, e, vi, vj, u) = e
update_vertex(gn, ē, vi, u) = vi
update_global(gn, ē, v̄, u) = u

Information propagation between different levels are achieved by aggregate functions. Three aggregate functions aggregate_neighbors, aggregate_edges and aggregate_vertices are defined to aggregate states.

GN block is realized into a abstract type GraphNet. User can make a subtype of GraphNet to customize GN block. Thus, a GN block is defined as a layer in GNN. MessagePassing is a subtype of GraphNet.

Update functions

update_edge acts as the first update function to apply to edge states. It takes edge state e, node i state vi, node j state vj and global state u. It is expected to return a feature vector for new edge state. update_vertex updates nodes state by taking aggregated edge state ē, node i state vi and global state u. It is expected to return a feature vector for new node state. update_global updates global state with aggregated information from edge and node. It takes aggregated edge state ē, aggregated node state v̄ and global state u. It is expected to return a feature vector for new global state. User can define their own behavior by overriding update functions.

GeometricFlux.update_edge — Function

update_edge(gn, e, vi, vj, u)

Update function for edge feature in graph network.

Arguments

gn::GraphNet: A graph network layer.

e: Edge feature.

vi: Node feature for node i.

vj: Node feature for neighbors of node i.

u: Global feature.

See also update_vertex, update_global, update_batch_edge, update_batch_vertex, aggregate_neighbors, aggregate_edges, aggregate_vertices.

GeometricFlux.update_vertex — Function

update_vertex(gn, ē, vi, u)

Update function for node feature in graph network.

Arguments

gn::GraphNet: A graph network layer.

ē: Aggregated edge feature.

vi: Node feature for node i.

u: Global feature.

See also update_edge, update_global, update_batch_edge, update_batch_vertex, aggregate_neighbors, aggregate_edges, aggregate_vertices.

GeometricFlux.update_global — Function

update_global(gn, ē, v̄, u)

Update function for global feature in graph network.

Arguments

gn::GraphNet: A graph network layer.

ē: Aggregated edge feature.

v̄: Aggregated node feature for node i.

u: Global feature.

See also update_edge, update_vertex, update_batch_edge, update_batch_vertex, aggregate_neighbors, aggregate_edges, aggregate_vertices.

GeometricFlux.update_batch_edge — Function

update_batch_edge(gn, el, E, V, u)

Returns new edge features of size (E_out_dim, #E, [batch_size]).

Arguments

gn::GraphNet: A graph network layer.

el::NamedTuple: Collection of graph information.

E: All edge features. Its size should be (E_in_dim, #E, [batch_size]).

V: All node features.

u: Global features.

See also update_edge, update_vertex, update_global, update_batch_vertex, aggregate_neighbors, aggregate_edges, aggregate_vertices.

GeometricFlux.update_batch_vertex — Function

update_batch_vertex(gn, el, Ē, V, u)

Returns new node features of size (V_out_dim, #V, [batch_size]).

Arguments

gn::GraphNet: A graph network layer.

el::NamedTuple: Collection of graph information.

Ē: All edge features. Its size should be (E_in_dim, #V, [batch_size]).

V: All node features. Its size should be (V_in_dim, #V, [batch_size]).

u: Global features.

See also update_edge, update_vertex, update_global, update_batch_edge, aggregate_neighbors, aggregate_edges, aggregate_vertices.

Aggregate functions

An aggregate function aggregate_neighbors aggregates edge states for edges incident to some node i into node-level information. Aggregate function aggregate_edges aggregates all edge states into global-level information. The last aggregate function aggregate_vertices aggregates all vertex states into global-level information.

GeometricFlux.aggregate_neighbors — Function

aggregate_neighbors(gn, el, aggr, E)

Returns aggregated neighbor features of size (E_out_dim, #V, [batch_size]).

Arguments

gn::GraphNet: A graph network layer.

el::NamedTuple: Collection of graph information.

aggr: Aggregate function to apply on neighbor features.

E: All edge features from neighbors. Its size should be (E_out_dim, #E, [batch_size]).

See also update_edge, update_vertex, update_global, update_batch_edge, update_batch_vertex, aggregate_edges, aggregate_vertices.

GeometricFlux.aggregate_edges — Function

aggregate_edges(gn, aggr, E)

Returns aggregated edge features of size (E_out_dim, 1, [batch_size]) for updating global feature.

Arguments

gn::GraphNet: A graph network layer.

aggr: Aggregate function to apply on edge features.

E: All edge features. Its size should be (E_out_dim, #E, [batch_size]).

See also update_edge, update_vertex, update_global, update_batch_edge, update_batch_vertex, aggregate_neighbors, aggregate_vertices.

GeometricFlux.aggregate_vertices — Function

aggregate_vertices(gn, aggr, V)

Returns aggregated node features of size (V_out_dim, 1, [batch_size]) for updating global feature.

Arguments

gn::GraphNet: A graph network layer.

aggr: Aggregate function to apply on node features.

V: All node features. Its size should be (V_out_dim, #V, [batch_size]).

See also update_edge, update_vertex, update_global, update_batch_edge, update_batch_vertex, aggregate_neighbors, aggregate_edges.

It is available for assigning aggregate function by assigning aggregate operations to propagate function.

propagate(gn, fg::FeaturedGraph, naggr=nothing, eaggr=nothing, vaggr=nothing)

naggr, eaggr and vaggr are arguments for aggregate_neighbors, aggregate_edges and aggregate_vertices, respectively. Available aggregate functions are assigned by following symbols to them: :add, :sub, :mul, :div, :max, :min and :mean.