FeaturedGraph

GraphSignals.FeaturedGraph — Type

FeaturedGraph(g, [mt]; directed=:auto, nf, ef, gf, T, N, E)

A type representing a graph structure and storing also arrays that contain features associated to nodes, edges, and the whole graph.

A FeaturedGraph can be constructed out of different objects g representing the connections inside the graph. When constructed from another featured graph fg, the internal graph representation is preserved and shared.

Arguments

g: Data representing the graph topology. Possible type are

An adjacency matrix.

An adjacency list.

A Graphs' graph, i.e. SimpleGraph, SimpleDiGraph from Graphs, or SimpleWeightedGraph, SimpleWeightedDiGraph from SimpleWeightedGraphs.

An AbstractFeaturedGraph object.

mt::Symbol: Matrix type for g in matrix form. if graph is in matrix form, mt is recorded as one of :adjm, :normedadjm, :laplacian, :normalized or :scaled.

directed: It specify that direction of a graph. It can be :auto, :directed and :undirected. Default value is :auto, which infers direction automatically.

nf: Node features.

ef: Edge features.

gf: Global features.

T: It specifies the element type of graph. Default value is the element type of g.

N: Number of nodes for g.

E: Number of edges for g.

Usage

using GraphSignals, CUDA

# Construct from adjacency list representation
g = [[2,3], [1,4,5], [1], [2,5], [2,4]]
fg = FeaturedGraph(g)

# Number of nodes and edges
nv(fg)  # 5
ne(fg)  # 10

# From a Graphs' graph
fg = FeaturedGraph(erdos_renyi(100, 20))

# Copy featured graph while also adding node features
fg = FeaturedGraph(fg, nf=rand(100, 5))

# Send to gpu
fg = fg |> cu

See also graph, node_feature, edge_feature, and global_feature.

GraphSignals.graph — Function

graph(fg)

Get referenced graph in fg.

Arguments

fg::AbstractFeaturedGraph: A concrete object of AbstractFeaturedGraph type.

GraphSignals.node_feature — Function

node_feature(fg)

Get node feature attached to fg.

Arguments

fg::AbstractFeaturedGraph: A concrete object of AbstractFeaturedGraph type.

GraphSignals.has_node_feature — Function

has_node_feature(::AbstractFeaturedGraph)

Check if node_feature is available or not for fg.

Arguments

fg::AbstractFeaturedGraph: A concrete object of AbstractFeaturedGraph type.

GraphSignals.edge_feature — Function

edge_feature(fg)

Get edge feature attached to fg.

Arguments

fg::AbstractFeaturedGraph: A concrete object of AbstractFeaturedGraph type.

GraphSignals.has_edge_feature — Function

has_edge_feature(::AbstractFeaturedGraph)

Check if edge_feature is available or not for fg.

Arguments

fg::AbstractFeaturedGraph: A concrete object of AbstractFeaturedGraph type.

GraphSignals.global_feature — Function

global_feature(fg)

Get global feature attached to fg.

Arguments

fg::AbstractFeaturedGraph: A concrete object of AbstractFeaturedGraph type.

GraphSignals.has_global_feature — Function

has_global_feature(::AbstractFeaturedGraph)

Check if global_feature is available or not for fg.

Arguments

fg::AbstractFeaturedGraph: A concrete object of AbstractFeaturedGraph type.

GraphSignals.subgraph — Function

subgraph(fg, nodes)

Returns a subgraph of type FeaturedSubgraph from a given featured graph fg. It constructs a subgraph by reserving nodes in a graph.

Arguments

fg::AbstractFeaturedGraph: A base featured graph to construct a subgraph.

nodes::AbstractVector: It specifies nodes to be reserved from fg.

GraphSignals.ConcreteFeaturedGraph — Function

ConcreteFeaturedGraph(fg; kwargs...)

This is a syntax sugar for construction for FeaturedGraph and FeaturedSubgraph object. It is an idempotent operation, which gives the same type of object as inputs. It wraps input fg again but reconfigures with kwargs.

Arguments

fg: FeaturedGraph and FeaturedSubgraph object.

Usage

julia> using GraphSignals

julia> adjm = [0 1 1 1;
               1 0 1 0;
               1 1 0 1;
               1 0 1 0];

julia> nf = rand(10, 4);

julia> fg = FeaturedGraph(adjm; nf=nf)
FeaturedGraph:
	Undirected graph with (#V=4, #E=5) in adjacency matrix
	Node feature:	ℝ^10 <Matrix{Float64}>

julia> ConcreteFeaturedGraph(fg, nf=rand(7, 4))
FeaturedGraph:
    Undirected graph with (#V=4, #E=5) in adjacency matrix
    Node feature:	ℝ^7 <Matrix{Float64}>