Convolution Layers

Graph Convolutional Layer

\[X' = \sigma(\hat{D}^{-1/2} \hat{A} \hat{D}^{-1/2} X \Theta)\]

where $\hat{A} = A + I$, $A$ denotes the adjacency matrix, and $\hat{D} = [\hat{d}_{ij}] = \sum_{j=0} [\hat{a}_{ij}]$ is degree matrix.

GeometricFlux.GCNConv — Type

GCNConv([graph, ]in=>out)
GCNConv([graph, ]in=>out, σ)

Graph convolutional layer.

Arguments

graph: should be a adjacency matrix, SimpleGraph, SimpleDiGraph (from LightGraphs)

or SimpleWeightedGraph, SimpleWeightedDiGraph (from SimpleWeightedGraphs). Is optionnal so you can give a FeaturedGraph to the layer instead of only the features.

in: the dimension of input features.

out: the dimension of output features.

bias::Bool=true: keyword argument, whether to learn the additive bias.

Data should be stored in (# features, # nodes) order. For example, a 1000-node graph each node of which poses 100 features is constructed. The input data would be a 1000×100 array.

Reference: Semi-supervised Classification with Graph Convolutional Networks

Chebyshev Spectral Graph Convolutional Layer

\[X' = \sum^{K-1}_{k=0} Z^{(k)} \Theta^{(k)}\]

where $Z^{(k)}$ is the $k$-th term of Chebyshev polynomials, and can be calculated by the following recursive form:

\[Z^{(0)} = X \\ Z^{(1)} = \hat{L} X \\ Z^{(k)} = 2 \hat{L} Z^{(k-1)} - Z^{(k-2)}\]

and $\hat{L} = \frac{2}{\lambda_{max}} L - I$.

GeometricFlux.ChebConv — Type

ChebConv([graph, ]in=>out, k)

Chebyshev spectral graph convolutional layer.

Arguments

graph: should be a adjacency matrix, SimpleGraph, SimpleDiGraph (from LightGraphs) or SimpleWeightedGraph,

SimpleWeightedDiGraph (from SimpleWeightedGraphs). Is optionnal so you can give a FeaturedGraph to the layer instead of only the features.

in: the dimension of input features.

out: the dimension of output features.

k: the order of Chebyshev polynomial.

bias::Bool=true: keyword argument, whether to learn the additive bias.

Reference: Convolutional Neural Networks on Graphs with Fast Localized Spectral Filtering

Graph Neural Network Layer

\[\textbf{x}_i' = \Theta_1 \textbf{x}_i + \sum_{j \in \mathcal{N}(i)} \Theta_2 \textbf{x}_j\]

GeometricFlux.GraphConv — Type

GraphConv([graph, ]in=>out)
GraphConv([graph, ]in=>out, aggr)

Graph neural network layer.

Arguments

graph: should be a adjacency matrix, SimpleGraph, SimpleDiGraph (from LightGraphs) or SimpleWeightedGraph,

SimpleWeightedDiGraph (from SimpleWeightedGraphs). Is optionnal so you can give a FeaturedGraph to the layer instead of only the features.

in: the dimension of input features.

out: the dimension of output features.

bias::Bool=true: keyword argument, whether to learn the additive bias.

aggr::Symbol=:add: an aggregate function applied to the result of message function. :add, :max and :mean are available.

Reference: Weisfeiler and Leman Go Neural: Higher-order Graph Neural Networks

Graph Attentional Layer

\[\textbf{x}_i' = \alpha_{i,i} \Theta \textbf{x}_i + \sum_{j \in \mathcal{N}(i)} \alpha_{i,j} \Theta \textbf{x}_j\]

where the attention coefficient $\alpha_{i,j}$ can be calculated from

\[\alpha_{i,j} = \frac{exp(LeakyReLU(\textbf{a}^T [\Theta \textbf{x}_i || \Theta \textbf{x}_j]))}{\sum_{k \in \mathcal{N}(i) \cup \{i\}} exp(LeakyReLU(\textbf{a}^T [\Theta \textbf{x}_i || \Theta \textbf{x}_k]))}\]

GeometricFlux.GATConv — Type

GATConv([graph, ]in=>out)

Graph attentional layer.

Arguments

graph: should be a adjacency matrix, SimpleGraph, SimpleDiGraph (from LightGraphs) or SimpleWeightedGraph,

SimpleWeightedDiGraph (from SimpleWeightedGraphs). Is optionnal so you can give a FeaturedGraph to the layer instead of only the features.

in: the dimension of input features.

out: the dimension of output features.

bias::Bool=true: keyword argument, whether to learn the additive bias.

negative_slope::Real=0.2: keyword argument, the parameter of LeakyReLU.

Reference: Graph Attention Networks

Gated Graph Convolution Layer

\[\textbf{h}^{(0)}_i = \textbf{x}_i || \textbf{0} \\ \textbf{h}^{(l)}_i = GRU(\textbf{h}^{(l-1)}_i, \sum_{j \in \mathcal{N}(i)} \Theta \textbf{h}^{(l-1)}_j)\]

where $\textbf{h}^{(l)}_i$ denotes the $l$-th hidden variables passing through GRU. The dimension of input $\textbf{x}_i$ needs to be less or equal to out.

GeometricFlux.GatedGraphConv — Type

GatedGraphConv([graph, ]out, num_layers)

Gated graph convolution layer.

Arguments

graph: should be a adjacency matrix, SimpleGraph, SimpleDiGraph (from LightGraphs) or SimpleWeightedGraph,

SimpleWeightedDiGraph (from SimpleWeightedGraphs). Is optionnal so you can give a FeaturedGraph to the layer instead of only the features.

out: the dimension of output features.

num_layers specifies the number of gated recurrent unit.

aggr::Symbol=:add: an aggregate function applied to the result of message function. :add, :max and :mean are available.

Reference: Gated Graph Sequence Neural Networks

Edge Convolutional Layer

\[\textbf{x}_i' = \sum_{j \in \mathcal{N}(i)} f_{\Theta}(\textbf{x}_i || \textbf{x}_j - \textbf{x}_i)\]

where $f_{\Theta}$ denotes a neural network parametrized by $\Theta$, i.e., a MLP.

GeometricFlux.EdgeConv — Type

EdgeConv(graph, nn)
EdgeConv(graph, nn, aggr)

Edge convolutional layer.

Arguments

graph: should be a adjacency matrix, SimpleGraph, SimpleDiGraph (from LightGraphs) or SimpleWeightedGraph, SimpleWeightedDiGraph (from SimpleWeightedGraphs).

nn: a neural network

aggr::Symbol=:max: an aggregate function applied to the result of message function. :add, :max and :mean are available.

Reference: Dynamic Graph CNN for Learning on Point Clouds