Group Convolutional Layers

$E(n)$-equivariant Convolutional Layer

It employs message-passing scheme and can be defined by following functions:

message function (Eq. 3 from paper): $m_{ij} = \phi_e(h_i^l, h_j^l, ||x_i^l - x_j^l||^2, a_{ij})$

aggregate (Eq. 5 from paper): $m_i = \sum_j m_{ij}$

update function (Eq. 6 from paper): $h_i^{l+1} = \phi_h(h_i^l, m_i)$

where $h_i^l$ and $h_j^l$ denotes the node feature for node $i$ and $j$, respectively, in $l$-th layer, as well as $x_i^l$ and $x_j^l$ denote the positional feature for node $i$ and $j$, respectively, in $l$-th layer. $a_{ij}$ is the edge feature for edge $(i,j)$. $\phi_e$ and $\phi_h$ are neural network for edges and nodes.

GeometricFlux.EEquivGraphConv — Type

EEquivGraphConv(in_dim=>out_dim, pos_dim, edge_dim; init=glorot_uniform)

E(n)-equivariant graph neural network layer.

Arguments

in_dim::Int: node feature dimension. Data is assumed to be of the form [feature; coordinate], so in_dim must strictly be less than the dimension of the input vectors.

out_dim: the output of the layer will have dimension out_dim + (dimension of input vector - in_dim).

hidden_dim::Int: dimension of positional encoding.

edge_dim::Int: dimension of edge feature.

init: neural network initialization function.

Examples

julia> in_dim, out_dim, pos_dim = 3, 5, 2
(3, 5, 2)

julia> egnn = EEquivGraphConv(in_dim=>out_dim, pos_dim, in_dim)
EEquivGraphConv(ϕ_edge=Chain(Dense(10 => 2), Dense(2 => 2)), ϕ_x=Chain(Dense(2 => 2), Dense(2 => 1; bias=false)), ϕ_h=Chain(Dense(5 => 2), Dense(2 => 5)))

See also WithGraph for training layer with static graph and EEquivGraphPE for positional encoding.

Reference: Victor Garcia Satorras, Emiel Hoogeboom, Max Welling (2021)