Linear Algebra

GeometricFlux.degrees — Function

degrees(g[, T; dir=:out])

Degree of each vertex. Return a vector which contains the degree of each vertex in graph g.

Arguments

g: should be a adjacency matrix, SimpleGraph, SimpleDiGraph (from LightGraphs) or SimpleWeightedGraph, SimpleWeightedDiGraph (from SimpleWeightedGraphs).

T: result element type of degree vector; default is the element type of g (optional).

dir: direction of degree; should be :in, :out, or :both (optional).

Examples

julia> using GeometricFlux

julia> m = [0 1 1; 1 0 0; 1 0 0];

julia> GeometricFlux.degrees(m)
3-element Array{Int64,1}:
 2
 1
 1

GeometricFlux.degree_matrix — Function

degree_matrix(g[, T; dir=:out])

Degree matrix of graph g. Return a matrix which contains degrees of each vertex in its diagonal. The values other than diagonal are zeros.

Arguments

g: should be a adjacency matrix, FeaturedGraph, SimpleGraph, SimpleDiGraph (from LightGraphs) or SimpleWeightedGraph, SimpleWeightedDiGraph (from SimpleWeightedGraphs).

T: result element type of degree vector; default is the element type of g (optional).

dir: direction of degree; should be :in, :out, or :both (optional).

Examples

julia> using GeometricFlux

julia> m = [0 1 1; 1 0 0; 1 0 0];

julia> GeometricFlux.degree_matrix(m)
3×3 SparseArrays.SparseMatrixCSC{Int64,Int64} with 3 stored entries:
  [1, 1]  =  2
  [2, 2]  =  1
  [3, 3]  =  1

GeometricFlux.inv_sqrt_degree_matrix — Function

inv_sqrt_degree_matrix(g[, T; dir=:out])

Inverse squared degree matrix of graph g. Return a matrix which contains inverse squared degrees of each vertex in its diagonal. The values other than diagonal are zeros.

Arguments

g: should be a adjacency matrix, FeaturedGraph, SimpleGraph, SimpleDiGraph (from LightGraphs) or SimpleWeightedGraph, SimpleWeightedDiGraph (from SimpleWeightedGraphs).

T: result element type of degree vector; default is the element type of g (optional).

dir: direction of degree; should be :in, :out, or :both (optional).

GeometricFlux.laplacian_matrix — Function

laplacian_matrix(g[, T; dir=:out])

Laplacian matrix of graph g.

Arguments

g: should be a adjacency matrix, FeaturedGraph, SimpleGraph, SimpleDiGraph (from LightGraphs) or SimpleWeightedGraph, SimpleWeightedDiGraph (from SimpleWeightedGraphs).

T: result element type of degree vector; default is the element type of g (optional).

dir: direction of degree; should be :in, :out, or :both (optional).

GeometricFlux.normalized_laplacian — Function

normalized_laplacian(g[, T; selfloop=false])

Normalized Laplacian matrix of graph g.

Arguments

g: should be a adjacency matrix, FeaturedGraph, SimpleGraph, SimpleDiGraph (from LightGraphs) or SimpleWeightedGraph, SimpleWeightedDiGraph (from SimpleWeightedGraphs).

T: result element type of degree vector; default is the element type of g (optional).

selfloop: adding self loop while calculating the matrix (optional).