Pooling layers

GeometricFlux.GlobalPool — Type

GlobalPool(aggr, dim...)

Global pooling layer.

It pools all features with aggr operation.

Arguments

aggr: An aggregate function applied to pool all features.

GeometricFlux.LocalPool — Type

LocalPool(aggr, cluster)

Local pooling layer.

It pools features with aggr operation accroding to cluster. It is implemented with scatter operation.

Arguments

aggr: An aggregate function applied to pool all features.

cluster: An index structure which indicates what features to aggregate with.

GeometricFlux.TopKPool — Type

TopKPool(adj, k, in_channel)

Top-k pooling layer.

Arguments

adj: Adjacency matrix of a graph.

k: Top-k nodes are selected to pool together.

in_channel: The dimension of input channel.