Models

Autoencoders

Graph Autoencoder

\[Z = enc(X, A) \\ \hat{A} = \sigma (ZZ^T)\]

where $A$ denotes the adjacency matrix.

GeometricFlux.GAE — Type

GAE(enc[, σ])

Graph autoencoder.

Arguments

enc: encoder. It can be any graph convolutional layer.

Encoder is specified by user and decoder will be InnerProductDecoder layer.

Reference: Variational Graph Auto-Encoders

Variational Graph Autoencoder

\[H = enc(X, A) \\ Z_{\mu}, Z_{logσ} = GCN_{\mu}(H, A), GCN_{\sigma}(H, A) \\ \hat{A} = \sigma (ZZ^T)\]

where $A$ denotes the adjacency matrix, $X$ denotes node features.

GeometricFlux.VGAE — Type

VGAE(enc[, σ])

Variational graph autoencoder.

Arguments

enc: encoder. It can be any graph convolutional layer.

Encoder is specified by user and decoder will be InnerProductDecoder layer.

Reference: Variational Graph Auto-Encoders

Special Layers

Inner-product Decoder

\[\hat{A} = \sigma (ZZ^T)\]

where $Z$ denotes the input matrix from encoder.

GeometricFlux.InnerProductDecoder — Type

InnerProductDecoder(σ)

Inner-product decoder layer.

Arguments

σ: activation function.

Reference: Variational Graph Auto-Encoders

Variational Encoder

\[H = enc(X, A) \\ Z_{\mu}, Z_{logσ} = GCN_{\mu}(H, A), GCN_{\sigma}(H, A)\]

GeometricFlux.VariationalEncoder — Type

VariationalEncoder(nn, h_dim, z_dim)

Variational encoder layer.

Arguments

nn: neural network. It can be any graph convolutional layer.

h_dim: dimension of hidden layer. This should fit the output dimension of nn.

z_dim: dimension of latent variable layer. This will be parametrized into μ and logσ.

Encoder can be any graph convolutional layer.

Reference: Variational Graph Auto-Encoders